pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "project_kit"
version = "0.0.1"
readme = "README.md"
dependencies = ["pyyaml"]

[tool.setuptools]
packages = [
	"project_kit",
	# "project_kit.utils"
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-64", "osx-arm64"]

[tool.pixi.pypi-dependencies]
project_kit = { path = ".", editable = true }

[tool.pixi.dependencies]
pycodestyle = ">=2.14.0,<3"
ipykernel = ">=6.30.1,<7"
jupyterlab = ">=4.4.7,<5"
numpy = ">=2.3.3,<3"
pandas = ">=2.3.2,<3"
